release_helpers/requirements.txt
PyYAML >= 6.0
Jinja2==3.1.2
docopt==0.6.2
MarkupSafe == 2.1.1
mdsplit==0.4.2